feat: minimum requirements completed

+++ server/DBManager.js
@@ -40,7 +40,7 @@ exports.getAllCars = function () {
                 return;
             }
             const cars = rows.map((row) =>
-                ({ id: row.id, brand: row.brand, category: row.category, year: row.year, fuel: row.fuel })
+                ({ id: row.id, brand: row.brand, category: row.category, year: row.year, fuel: row.fuel, model: row.model })
             );
             resolve(cars);
         });
@@ -48,12 +48,12 @@ exports.getAllCars = function () {
 }
 
 exports.getByBrand = function (brand) {
-    const brands=brand.split(',');
+    const brands = brand.split(',');
     let params = "?";
     if (brands.length > 1) {
-        for(let i = 1; i < brands.length; i++) {
-            params=params.concat(", ?")
-          }
+        for (let i = 1; i < brands.length; i++) {
+            params = params.concat(", ?")
+        }
     }
     return new Promise((resolve, reject) => {
         const sql = `SELECT *  FROM cars where brand IN (${params}) ;`
@@ -63,7 +63,7 @@ exports.getByBrand = function (brand) {
                 return;
             }
             const cars = rows.map((row) =>
-                ({ id: row.id, brand: row.brand, category: row.category, year: row.year, fuel: row.fuel })
+                ({ id: row.id, brand: row.brand, category: row.category, year: row.year, fuel: row.fuel, model: row.model })
             );
             resolve(cars);
         });
@@ -71,19 +71,19 @@ exports.getByBrand = function (brand) {
 }
 
 exports.getByCategoryAndBrand = function (category, brand) {
-    const categories=category.split(',');
+    const categories = category.split(',');
     let params = "?";
     if (categories.length > 1) {
-        for(let i = 1; i < categories.length; i++) {
-            params=params.concat(", ?")
-          }
+        for (let i = 1; i < categories.length; i++) {
+            params = params.concat(", ?")
+        }
     }
-    const brands=brand.split(',');
+    const brands = brand.split(',');
     let params2 = "?";
     if (brands.length > 1) {
-        for(let i = 1; i < brands.length; i++) {
-            params2=params2.concat(", ?")
-          }
+        for (let i = 1; i < brands.length; i++) {
+            params2 = params2.concat(", ?")
+        }
     }
     return new Promise((resolve, reject) => {
         const sql = `SELECT *  FROM cars where brand IN (${params2}) AND category IN  (${params});`
@@ -94,7 +94,7 @@ exports.getByCategoryAndBrand = function (category, brand) {
                 return;
             }
             const cars = rows.map((row) =>
-                ({ id: row.id, brand: row.brand, category: row.category, year: row.year, fuel: row.fuel })
+                ({ id: row.id, brand: row.brand, category: row.category, year: row.year, fuel: row.fuel, model: row.model })
             );
             resolve(cars);
         });
@@ -102,12 +102,12 @@ exports.getByCategoryAndBrand = function (category, brand) {
 }
 
 exports.getByCategory = function (category) {
-    const categories=category.split(',');
+    const categories = category.split(',');
     let params = "?";
     if (categories.length > 1) {
-        for(let i = 1; i < categories.length; i++) {
-            params=params.concat(", ?")
-          }
+        for (let i = 1; i < categories.length; i++) {
+            params = params.concat(", ?")
+        }
     }
     return new Promise((resolve, reject) => {
         const sql = `SELECT *  FROM cars where category IN (${params});`
@@ -117,7 +117,7 @@ exports.getByCategory = function (category) {
                 return;
             }
             const cars = rows.map((row) =>
-                ({ id: row.id, brand: row.brand, category: row.category, year: row.year, fuel: row.fuel })
+                ({ id: row.id, brand: row.brand, category: row.category, year: row.year, fuel: row.fuel, model: row.model })
             );
             resolve(cars);
         });
@@ -126,24 +126,24 @@ exports.getByCategory = function (category) {
 
 exports.getAllBookings = function (userID) {
     return new Promise((resolve, reject) => {
-        const sql = "SELECT *  FROM bookings where userId=?;"
+        const sql = "SELECT b.id, startDate, endDate, b.category, extraDrivers, estimation,insurance,age,price,carId, userId,brand, fuel,year,model  FROM bookings b , cars c where userId=? and b.carId=c.id;"
         db.all(sql, [userID], (err, rows) => {
             if (err) {
                 reject(err);
                 return;
             }
             const bookings = rows.map((row) =>
-                ({ id: row.id, startDate: row.startDate, endDate: row.endDate, age: row.age, category: row.category, extraDrivers: row.extraDrivers, estimation: row.estimation, insurance: row.insurance, price: price, carId: carId, userId: userId })
+                ({ id: row.id, startDate: row.startDate, endDate: row.endDate, age: row.age, category: row.category, extraDrivers: row.extraDrivers, estimation: row.estimation, insurance: row.insurance, price: row.price, carId: row.carId, userId: row.userId, fuel: row.fuel, brand: row.brand, year: row.year, model: row.model })
             );
             resolve(bookings);
         });
     });
 }
 
-exports.removeBooking = function (id = -1) {
+exports.removeBooking = function (id = -1, userID) {
     return new Promise((resolve, reject) => {
-        const sql = "DELETE  FROM bookings WHERE id=?"
-        db.run(sql, [id], (err) => {
+        const sql = "DELETE  FROM bookings WHERE id=? and userId=?"
+        db.run(sql, [id, userID], (err) => {
             if (err) {
                 reject(err);
                 return;
@@ -152,11 +152,29 @@ exports.removeBooking = function (id = -1) {
         })
     })
 }
+exports.bookCar = function (startDate, endDate, category) {
+    return new Promise((resolve, reject) => {
+        const subSql1 = "SELECT carId FROM bookings WHERE DATE(startDate)<=DATE(?) AND DATE(endDate)<=DATE(?) AND DATE(endDate)>=DATE(?)" // correct order -> start,end,start
+        const subSql2 = "SELECT carId FROM bookings WHERE DATE(startDate)>=DATE(?) AND DATE(endDate)<=DATE(?)" // correct order -> start,end
+        const subSql3 = "SELECT carId FROM bookings WHERE DATE(startDate)>=DATE(?) AND DATE(startDate)<=DATE(?) and DATE(endDate)>=DATE(?)" // correct order -> start,end,end
+        const subSql4 = "SELECT carId FROM bookings WHERE DATE(startDate)<=DATE(?) AND DATE(endDate)>=DATE(?)" // correct order -> start,end
+        const sql = `SELECT id FROM cars WHERE id NOT IN (${subSql1}) and id NOT IN(${subSql2}) and id NOT IN (${subSql3}) and id NOT IN (${subSql4}) AND category=?;`;
+        db.get(sql, [startDate, endDate, startDate, startDate, endDate, startDate, endDate, endDate,startDate, endDate, category], (err, row) => {
+            if (err) {
+                reject(err);
+                return;
+            }
+            if (row)
+                resolve(row.id);
+            else resolve(null);
+        })
+    })
+}
 
-exports.createBooking = function (booking, userID) {
+exports.createBooking = function (booking) {
     return new Promise((resolve, reject) => {
         const sql = "INSERT  INTO bookings(startDate,endDate,age,category,extraDrivers,estimation, insurance, price, carId, userId) VALUES(datetime(?),datetime(?),?,?,?,?,?,?,?,?)"
-        db.run(sql, [booking.startDate.toString(), booking.endDate.toString(), booking.age, booking.category, booking.extraDrivers, booking.estimation, booking.insurance, booking.price, booking.carId, userID], (err) => {
+        db.run(sql, [booking.startDate.toString(), booking.endDate.toString(), booking.age, booking.category, booking.extraDrivers, booking.estimation, booking.insurance, booking.price, booking.carId, booking.userId], (err) => {
             if (err) {
                 reject(err);
                 return;
@@ -200,6 +218,19 @@ exports.getUser = function (id = -1) {
     });
 }
 
+exports.updateUserBookings = function (id = -1, sum) {
+    return new Promise((resolve, reject) => {
+        const sql = "UPDATE USERS  SET nbookings=nbookings+? WHERE id=?;"
+        db.run(sql, [sum, id], (err) => {
+            if (err) {
+                reject(err);
+                return;
+            }
+            resolve(true);
+        });
+    });
+}
+
 exports.countCars = function (category) {
     return new Promise((resolve, reject) => {
         const sql = `SELECT COUNT(*)  FROM cars where category=?;`
@@ -213,10 +244,15 @@ exports.countCars = function (category) {
     });
 }
 
-exports.getConcurrentBookings =function (startDate, endDate, category){
-    return new Promise((resolve, reject)=>{
-        const sql = "SELECT * FROM bookings WHERE (DATE(startDate)<DATE(?) AND DATE(endDate)<DATE(?) AND DATE(endDate)>DATE(?) OR DATE(startDate)>DATE(?) AND DATE(endDate)<DATE(?) OR DATE(startDate)>DATE(?) AND DATE(startDate)<DATE(?) and DATE(endDate)>DATE(?)) AND category=? ;"
-        db.all(sql, [startDate,endDate,startDate,startDate,endDate,startDate,endDate,endDate,category], (err, rows) => {
+exports.getConcurrentBookings = function (startDate, endDate, category) {
+    console.log(startDate);
+    return new Promise((resolve, reject) => {
+        const subSql1 = "SELECT id FROM bookings WHERE DATE(startDate)<=DATE(?) AND DATE(endDate)<=DATE(?) AND DATE(endDate)>=DATE(?)" // correct order -> start,end,start
+        const subSql2 = "SELECT id FROM bookings WHERE DATE(startDate)>=DATE(?) AND DATE(endDate)<=DATE(?)" // correct order -> start,end
+        const subSql3 = "SELECT id FROM bookings WHERE DATE(startDate)>=DATE(?) AND DATE(startDate)<=DATE(?) and DATE(endDate)>=DATE(?)" // correct order -> start,end,end
+        const subSql4 = "SELECT id FROM bookings WHERE DATE(startDate)<=DATE(?) AND DATE(endDate)>=DATE(?)" // correct order -> start,end
+        const sql = `SELECT * FROM bookings WHERE id IN(${subSql1}) OR id IN(${subSql2}) OR id IN(${subSql3}) OR id IN (${subSql4}) AND category=? ;`
+        db.all(sql, [startDate, endDate, startDate, startDate, endDate, startDate, endDate, endDate, startDate, endDate, category], (err, rows) => {
             if (err) {
                 reject(err);
                 return;
